호텔상세정보 객실예약페이지 연동 및 객실예약페이지 이벤트동작 및 리듀서추가

+++ src/containers/auth/loginForm.js
@@ -1,7 +1,8 @@
-import { useEffect, useState, useNavigate } from "react";
+import { useEffect, useState } from "react";
 import AuthLoginForm from "../../components/auth/authLoginForm";
 import { changeField, resetField, initializeForm, login } from "../../modules/auth";
 import { useDispatch, useSelector } from "react-redux";
+import { useNavigate } from 'react-router-dom';
 /**
  * 
  * 로그인 컨테이너
@@ -82,15 +83,16 @@ const Login =() => {
         if(!mailCheck.test(form.email)){ setErrCount(1)}
         if(!pwCheck.test(form.password)){ setErrCount(2) }
         if(!mailCheck.test(form.email) && !pwCheck.test(form.password)){ setErrCount(3) }
-        console.log('aa');
+       
         if(mailCheck.test(form.email) && pwCheck.test(form.password)){//정규식통과시 로그인 api연동
+
              //console.log(form.email, form.password, form.role);
              const { email,  password, role } = form;
+             console.log(email, password, role);
              dispatch(login({ email, password, role }));
-             console.log('bb');
         }
     }
-    //const navigate = useNavigate();
+    const navigate = useNavigate();
 
      //로그인성공/실패처리
      useEffect(() => {
@@ -106,11 +108,23 @@ const Login =() => {
             console.log(auth);
             //navigate('/');
             //window.location.href = "./auth/token/login";
-            localStorage.clear();
-            localStorage.setItem('accessToken',auth.headers.authorization);
-            localStorage.setItem('refreshToken',auth.headers.refreshtoken);
-            localStorage.setItem('role',auth.data.role);
-            localStorage.setItem('email',auth.data.email);
+            if(auth.data.result === 'OK'){
+                localStorage.clear();
+                localStorage.setItem('accessToken',auth.headers.authorization);
+                localStorage.setItem('refreshToken',auth.headers.refreshtoken);
+                localStorage.setItem('role',auth.data.role);
+                localStorage.setItem('email',auth.data.email);
+                console.log(auth.headers.authorization,auth.data.role, auth.data.email);
+                if(auth.data.role === 2){
+                    navigate('/auth/hotel/main');
+                }else{
+                    navigate('/');
+                }
+            }else{
+                console.log('아이디 비밀번호 error', auth);
+                alert('아이도 또는 비밀번호가 틀립니다. 다시 입력해주세요.');
+            }
+            
             //console.log(auth.headers.authorization,auth.data.role, auth.data.email);
             //console.log('refresh',auth.headers.authorization);
             //console.log(localStorage.getItem('accessToken'),localStorage.getItem('refreshToken'),localStorage.getItem('role'),localStorage.getItem('email'));
